[#231] Refactor: Clean up all index.ts files from src
Cleaned up index.ts files from the following folders:
  - src/brokers
  - src/constants
  - src/i18n
  - src/shared-components
