+++ src/brokers/add-broker/AddBroker.component.tsx
@@ -9,7 +9,7 @@ import {
 } from '../../reducers/7.12/reducer';
 import { FormView } from '../../shared-components/FormView/FormView';
 import { YamlEditorView } from '../../shared-components/YamlEditorView/YamlEditorView';
-import { EditorToggle } from './components/EditorToggle';
+import { EditorToggle } from './components/EditorToggle/EditorToggle';
 
 type AddBrokerProps = {
   onCreateBroker: (data?: K8sResourceCommon) => void;
@@ -20,7 +20,7 @@ type AddBrokerProps = {
   isUpdate?: boolean;
 };
 
-const AddBroker: FC<AddBrokerProps> = ({
+export const AddBroker: FC<AddBrokerProps> = ({
   onCreateBroker,
   notification,
   isUpdate,
@@ -60,5 +60,3 @@ const AddBroker: FC<AddBrokerProps> = ({
     </>
   );
 };
-
-export { AddBroker };
