+++ src/brokers/add-broker/components/EditorToggle/EditorToggle.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Flex, Radio } from '@patternfly/react-core';
-import { useTranslation } from '../../../../i18n';
+import { useTranslation } from '../../../../i18n/i18n';
 import { EditorType } from '../../../../reducers/7.12/reducer';
 
 type EditorToggleProps = {
