--- src/brokers/add-broker/components/EditorToggle/index.ts
@@ -1,1 +1,0 @@
-export * from './EditorToggle';
