+++ src/brokers/broker-details/components/BrokerDetailsBreadcrumb/BrokerDetailsBreadcrumb.tsx
@@ -10,10 +10,10 @@ import {
   Level,
   LevelItem,
 } from '@patternfly/react-core';
-import { useTranslation } from '../../../../i18n';
+import { useTranslation } from '../../../../i18n/i18n';
 import { k8sDelete } from '@openshift-console/dynamic-plugin-sdk';
 import { AMQBrokerModel } from '../../../../k8s';
-import { PreConfirmDeleteModal } from '../../../view-brokers/components';
+import { PreConfirmDeleteModal } from '../../../view-brokers/components/PreConfirmDeleteModal/PreConfirmDeleteModal';
 import { useNavigate } from 'react-router-dom-v5-compat';
 
 export type BrokerDetailsBreadcrumbProps = {
