--- src/brokers/broker-details/components/BrokerDetailsBreadcrumb/index.ts
@@ -1,1 +1,0 @@
-export * from './BrokerDetailsBreadcrumb';
