--- src/brokers/broker-details/components/Clients/index.ts
@@ -1,1 +1,0 @@
-export * from './Clients.container';
