+++ src/brokers/broker-details/components/Configuration/Configuration.container.tsx
@@ -5,8 +5,8 @@ import {
   BrokerConditionTypes,
   K8sResourceCondition,
 } from '../../../../k8s';
-import { Loading } from '../../../../shared-components';
-import { useTranslation } from '../../../../i18n';
+import { Loading } from '../../../../shared-components/Loading/Loading';
+import { useTranslation } from '../../../../i18n/i18n';
 
 export type ConfigurationContainerProps = {
   configurationSettings: BrokerCR;
