--- src/brokers/broker-details/components/Configuration/index.ts
@@ -1,1 +1,0 @@
-export * from './Configuration.container';
