+++ src/brokers/broker-details/components/Overview/Metrics/components/ChartCPUUsage/ChartCPUUsage.tsx
@@ -17,9 +17,12 @@ import {
 } from '@openshift-console/dynamic-plugin-sdk';
 import { ChartSkeletonLoader } from '../ChartSkeletonLoader/ChartSkeletonLoader';
 import { EmptyStateNoMetricsData } from '../EmptyStateNoMetricsData/EmptyStateNoMetricsData';
-import { chartHeight, chartPadding } from '../../../../../../../constants';
+import {
+  chartHeight,
+  chartPadding,
+} from '../../../../../../../constants/constants';
 import { useChartWidth } from '../../hooks/useChartWidth';
-import { useTranslation } from '../../../../../../../i18n';
+import { useTranslation } from '../../../../../../../i18n/i18n';
 import {
   chartTheme,
   valueFormatter,
