--- src/brokers/broker-details/components/Overview/index.ts
@@ -1,1 +1,0 @@
-export * from './Overview.container';
