--- src/brokers/broker-details/components/index.ts
@@ -1,1 +1,0 @@
-export * from './Overview';
