+++ src/brokers/broker-pods/components/BrokerPodsBreadcrumb/BrokerPodsBreadcrumb.tsx
@@ -10,8 +10,8 @@ import {
   Level,
   LevelItem,
 } from '@patternfly/react-core';
-import { PreConfirmDeleteModal } from '../../../view-brokers/components/PreConfirmDeleteModal';
-import { useTranslation } from '../../../../i18n';
+import { PreConfirmDeleteModal } from '../../../view-brokers/components/PreConfirmDeleteModal/PreConfirmDeleteModal';
+import { useTranslation } from '../../../../i18n/i18n';
 import { k8sDelete } from '@openshift-console/dynamic-plugin-sdk';
 import { AMQBrokerModel } from '../../../../k8s';
 import { useNavigate } from 'react-router-dom-v5-compat';
