--- src/brokers/broker-pods/components/BrokerPodsBreadcrumb/index.ts
@@ -1,1 +1,0 @@
-export * from './BrokerPodsBreadcrumb';
