--- src/brokers/view-brokers/components/BrokersList/index.ts
@@ -1,1 +1,0 @@
-export * from './BrokersList';
