--- src/brokers/view-brokers/components/PreConfirmDeleteModal/index.ts
@@ -1,1 +1,0 @@
-export * from './PreConfirmDeleteModal';
