--- src/brokers/view-brokers/components/index.ts
@@ -1,2 +1,0 @@
-export * from './BrokersList';
-export * from './PreConfirmDeleteModal';
