--- src/i18n/index.ts
@@ -1,2 +1,0 @@
-export { default } from './i18nProvider';
-export * from './i18n';
