+++ src/shared-components/FormView/BrokerProperties/ConfigurationPage/AcceptorsConfigPage/AcceptorConfigSection/AcceptorConfigPage/AcceptorConfigPage.tsx
@@ -24,7 +24,7 @@ import {
 } from '../../../../../../../reducers/7.12/reducer';
 import { FC, useContext } from 'react';
 import { ConfigType } from '../../../ConfigurationPage';
-import { useTranslation } from '../../../../../../../i18n';
+import { useTranslation } from '../../../../../../../i18n/i18n';
 import { PresetAlertPopover } from './PresetAlertPopover/PresetAlertPopover';
 import { SelectExposeMode } from './SelectExposeMode/SelectExposeMode';
 import { OtherParameters } from './OtherParameters/OtherParameters';
