+++ src/shared-components/FormView/BrokerProperties/ConfigurationPage/AcceptorsConfigPage/AcceptorConfigSection/AcceptorConfigPage/ListPresets/ListPresets.tsx
@@ -4,8 +4,7 @@ import {
   FormGroup,
 } from '@patternfly/react-core';
 import { FC, useContext } from 'react';
-
-import { useTranslation } from '../../../../../../../../i18n';
+import { useTranslation } from '../../../../../../../../i18n/i18n';
 import { Acceptor } from '../../../../../../../../k8s';
 import {
   ArtemisReducerOperations,
