+++ src/shared-components/FormView/BrokerProperties/ConfigurationPage/AcceptorsConfigPage/AcceptorConfigSection/AcceptorConfigPage/PresetAlertPopover/PresetAlertPopover.tsx
@@ -5,7 +5,7 @@ import {
   getAcceptor,
   getCertManagerResourceTemplateFromAcceptor,
 } from '../../../../../../../../reducers/7.12/reducer';
-import { useTranslation } from '../../../../../../../../i18n';
+import { useTranslation } from '../../../../../../../../i18n/i18n';
 import { Alert, Popover } from '@patternfly/react-core';
 import { BellIcon, WarningTriangleIcon } from '@patternfly/react-icons';
