+++ src/shared-components/FormView/BrokerProperties/ConfigurationPage/AcceptorsConfigPage/AcceptorConfigSection/AcceptorConfigPage/SelectExposeMode/SelectExposeMode.tsx
@@ -1,7 +1,7 @@
 import { ConfigType } from '../../../../ConfigurationPage';
 
 import { FC, useState } from 'react';
-import { useTranslation } from '../../../../../../../../i18n';
+import { useTranslation } from '../../../../../../../../i18n/i18n';
 import { PresetAlertPopover } from '../PresetAlertPopover/PresetAlertPopover';
 import { ExposeMode } from '../../../../../../../../reducers/7.12/reducer';
 import {
