+++ src/shared-components/FormView/BrokerProperties/ConfigurationPage/AcceptorsConfigPage/AcceptorConfigSection/PresetButton/PresetButton.tsx
@@ -15,7 +15,7 @@ import {
 } from '@patternfly/react-core';
 import { CSSProperties, FC, useContext, useState } from 'react';
 
-import { useTranslation } from '../../../../../../../i18n';
+import { useTranslation } from '../../../../../../../i18n/i18n';
 import { Acceptor } from '../../../../../../../k8s';
 import {
   ArtemisReducerOperations,
