+++ src/shared-components/YamlEditorView/YamlEditorView.tsx
@@ -13,8 +13,8 @@ import {
   useAccessReview,
 } from '@openshift-console/dynamic-plugin-sdk';
 import { AMQBrokerModel, BrokerCR } from '../../k8s';
-import { Loading } from '../../shared-components';
-import { useTranslation } from '../../i18n';
+import { Loading } from '../../shared-components/Loading/Loading';
+import { useTranslation } from '../../i18n/i18n';
 import {
   ArtemisReducerOperations,
   BrokerCreationFormDispatch,
